test(e2e): mock realtime connections in client tests * avoid real Upstash connections during E2E tests * add connected and unavailable EventSource mocks * test polling only as realtime fallback * keep realtime-connected tests independent from Redis
