+++ apps/client-e2e/src/pages/overview.page.ts
@@ -107,62 +107,6 @@ export class OverviewPage {
     });
   }
 
-  async mockRealtimeConnected(): Promise<void> {
-    await this.page.addInitScript(() => {
-      class MockEventSource {
-        static readonly CONNECTING = 0;
-        static readonly OPEN = 1;
-        static readonly CLOSED = 2;
-
-        readonly CONNECTING = 0;
-        readonly OPEN = 1;
-        readonly CLOSED = 2;
-
-        readonly url: string;
-        readonly withCredentials = false;
-
-        readyState = MockEventSource.CONNECTING;
-
-        onopen: ((event: Event) => void) | null = null;
-        onmessage: ((event: MessageEvent<string>) => void) | null = null;
-        onerror: ((event: Event) => void) | null = null;
-
-        constructor(url: string | URL) {
-          this.url = url.toString();
-
-          setTimeout(() => {
-            this.readyState = MockEventSource.OPEN;
-            this.onopen?.(new Event('open'));
-          });
-        }
-
-        close(): void {
-          this.readyState = MockEventSource.CLOSED;
-        }
-
-        addEventListener(): void {
-          /* empty */
-        }
-
-        removeEventListener(): void {
-          /* empty */
-        }
-
-        dispatchEvent(): boolean {
-          return true;
-        }
-      }
-
-      window.EventSource = MockEventSource as unknown as typeof EventSource;
-    });
-  }
-
-  async mockRealtimeUnavailable(): Promise<void> {
-    await this.page.route('**/livestream', async (route) => {
-      await route.abort('connectionrefused');
-    });
-  }
-
   getFirstFixture(): Locator {
     return this.fixtureItems.first();
   }
